+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Custom Action - Fix the bad endpoints if some actions have the same name.
 
 ## RELEASE 0.1.29 - 2016-10-14
 ### Fixed